data_prep/df_requirements.txt
annotated-types==0.7.0
anyio==4.9.0
apache-beam==2.63.0
attrs==25.3.0
beam-postgres-connector==0.1.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==2.2.1
crcmod==1.7
cryptography==44.0.2
Deprecated==1.2.18
dill==0.3.1.1
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.16
dunamai==1.23.1
fastavro==1.10.0
fasteners==0.19
google-api-core==2.24.2
google-apitools==0.5.31
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.31.0
google-cloud-bigquery-storage==2.30.0
google-cloud-bigtable==2.30.0
google-cloud-core==2.4.3
google-cloud-datastore==2.20.2
google-cloud-dlp==3.29.0
google-cloud-language==2.17.1
google-cloud-pubsub==2.29.0
google-cloud-pubsublite==1.12.0
google-cloud-recommendations-ai==0.10.17
google-cloud-resource-manager==1.14.2
google-cloud-spanner==3.53.0
google-cloud-storage==2.19.0
google-cloud-videointelligence==2.16.1
google-cloud-vision==3.10.1
google-crc32c==1.7.1
google-genai==1.8.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.65.5
grpcio-status==1.65.5
h11==0.14.0
hdfs==2.7.3
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.6
jsonpickle==3.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
keyrings.google-artifactregistry-auth==1.1.2
MarkupSafe==3.0.2
more-itertools==10.6.0
numpy==2.2.4
oauth2client==4.1.3
objsize==0.7.1
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pluggy==1.5.0
poetry-dynamic-versioning==0.24.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pydot==1.4.2
pymongo==4.11.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
rsa==4.9
SecretStorage==3.3.3
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
sqlparse==0.5.3
tomlkit==0.13.2
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
websockets==15.0.1
wrapt==1.17.2
zipp==3.21.0
zstandard==0.23.0